+++ src/declare-modules.d.ts
@@ -1,7 +1,9 @@
-import { EdgeCorePlugins } from 'edge-core-js/types'
-
-declare global {
-  interface Window {
-    addEdgeCorePlugins?: (plugins: EdgeCorePlugins) => void
+declare module 'react-native' {
+  export const NativeModules: {
+    EdgeExchangePluginsModule: {
+      getConstants: () => {
+        sourceUri: string
+      }
+    }
   }
 }