+++ src/index.ts
@@ -1,14 +1,7 @@
-import { makeBitMaxPlugin } from './rate/bitmax'
-import { makeCoinbasePlugin } from './rate/coinbase'
-import { makeCoincapPlugin } from './rate/coincap'
-import { makeCoinGeckoPlugin } from './rate/coingecko'
-import { makeCoinmonitorPlugin } from './rate/coinmonitor'
-import { makeCompoundPlugin } from './rate/compound'
-import { makeConstantRatePlugin } from './rate/constantRate'
-import { makeCurrencyconverterapiPlugin } from './rate/currencyconverterapi'
-import { makeEdgeRatesPlugin } from './rate/edgeRates'
-import { makeNomicsPlugin } from './rate/nomics'
-import { makeWazirxPlugin } from './rate/wazirx'
+import 'regenerator-runtime/runtime'
+
+import type { EdgeCorePlugins } from 'edge-core-js/types'
+
 import { makeChangeHeroPlugin } from './swap/changehero'
 import { makeChangeNowPlugin } from './swap/changenow'
 import { makeLifiPlugin } from './swap/defi/lifi'
@@ -25,20 +18,7 @@ import { makeSideshiftPlugin } from './swap/sideshift'
 import { makeSwapuzPlugin } from './swap/swapuz'
 import { makeTransferPlugin } from './swap/transfer'
 
-const edgeCorePlugins = {
-  // Rate plugins:
-  bitmax: makeBitMaxPlugin,
-  coinbase: makeCoinbasePlugin,
-  coincap: makeCoincapPlugin,
-  coingecko: makeCoinGeckoPlugin,
-  coinmonitor: makeCoinmonitorPlugin,
-  compound: makeCompoundPlugin,
-  constantRate: makeConstantRatePlugin,
-  currencyconverterapi: makeCurrencyconverterapiPlugin,
-  edgeRates: makeEdgeRatesPlugin,
-  nomics: makeNomicsPlugin,
-  wazirx: makeWazirxPlugin,
-
+const plugins = {
   // Swap plugins:
   changehero: makeChangeHeroPlugin,
   changenow: makeChangeNowPlugin,
@@ -57,11 +37,17 @@ const edgeCorePlugins = {
   letsexchange: makeLetsExchangePlugin
 }
 
+declare global {
+  interface Window {
+    addEdgeCorePlugins?: (plugins: EdgeCorePlugins) => void
+  }
+}
+
 if (
   typeof window !== 'undefined' &&
   typeof window.addEdgeCorePlugins === 'function'
 ) {
-  window.addEdgeCorePlugins(edgeCorePlugins)
+  window.addEdgeCorePlugins(plugins)
 }
 
-export default edgeCorePlugins
+export default plugins