+++ src/swap/swapuz.ts
@@ -18,6 +18,7 @@ import {
   SwapCurrencyError
 } from 'edge-core-js/types'
 
+import { div18 } from '../util/biggystringplus'
 import {
   checkInvalidCodes,
   ensureInFuture,
@@ -27,8 +28,7 @@ import {
   isLikeKind,
   makeSwapPluginQuote,
   SwapOrder
-} from '../swap-helpers'
-import { div18 } from '../util/biggystringplus'
+} from '../util/swapHelpers'
 import { convertRequest, getAddress } from '../util/utils'
 import { EdgeSwapRequestPlugin } from './types'